Turned off the selective zero truncation for the local properties...
I think??? also tried to do it only for atomic charges, since it really
doesn't make sense there, while keeping the existing behaviour for vdW
parameters (volume) and electron binding energies

NOTE this was really tricky to merge with the new local property stuff
in turbogap, so ended up changing the way the zero-truncation flag is
stored. Hopefully this will do exactly the same thing as before, but it
needs to be checked.